Hi Ernest,
Here is the new script for the 2nd exercise I have worked on using the csv file. Please review it and let me know if there is anything I can improve on.

I am still trying to play around with and improve the legend and the text in the maps so that they are much clearer.

Thanks