Hi,
Here is the script which I have tried to work on for the spatialSampler exercise on the Malawi shapefiles data.

You will notice that I tried to do a subset of the regions and district but my arguments/commands seem not to be correct. I keep getting errors.

I am also having issues with the following:
   - calculation of the n_grids and it is
     affecting the rest of the subsequent
     steps.
   - subset village_locations to those that
     are only in the survey area
   - Selecting the nearest 1 village to the
     sampling points (here data.x =
     "...LONG", data.y = "....LAT".

Let me know what I need to rework on.

Thanks.